+++ app/javascript/components/help/CounterStackingSection.tsx
@@ -74,7 +74,7 @@ export default function CounterStackingSection() {
         <svg width={504} height={134} viewBox='0 0 720 192' style={{ minWidth: 504 }}>
           <MapCounterOverlay map={map} setOverlay={() => {}} selectionCallback={() => {}}
                              xx={0} yy={0} mapScale={1} shiftX={0} shiftY={44} maxX={0} maxY={0}
-                             counters={map.countersAt(new Coordinate(0,0))} scale={1}
+                             counters={map.countersAt(new Coordinate(0,0))} scale={1} updateCallback={() => {}}
                              svgRef={svgRef as React.MutableRefObject<HTMLElement>} />
           <path d={roundedRectangle(0,0,736,192,0)}
                 style={{ stroke: clearColor, strokeWidth: 0.5, fill: clearColor }}/>
@@ -89,7 +89,7 @@ export default function CounterStackingSection() {
         <svg width={627} height={134} viewBox='0 0 896 192' style={{ minWidth: 627 }}>
           <MapCounterOverlay map={map} setOverlay={() => {}} selectionCallback={() => {}}
                              xx={0} yy={0} mapScale={1} shiftX={0} shiftY={44} maxX={0} maxY={0}
-                             counters={map.countersAt(new Coordinate(1,1))} scale={1}
+                             counters={map.countersAt(new Coordinate(1,1))} scale={1} updateCallback={() => {}}
                              svgRef={svgRef as React.MutableRefObject<HTMLElement>} />
           <path d={roundedRectangle(0,0,896,192,0)}
                 style={{ stroke: clearColor, strokeWidth: 0.5, fill: clearColor }}/>